requirements.txt
Pillow
pydicom
numpy
sqlalchemy
psycopg2-binary
scikit-learn
opencv-python
